test(pkg/gw): enhance immugw handlers tests to reuse buffconn server and use in memory file cache
